intercom and register working
